Moves `validateResourceSchemas` to testing
